fix(analyzer): widen types and propagate nullability instead of guessing
Eight issues here shared one root: nothing in the analyzer widened. Type
resolution took the first or leftmost operand and kept it, so int + numeric
stayed int, and a function's result was assumed NOT NULL regardless of its
arguments. Every one of these is a silent wrong answer -- the generated code
compiles and then hands back a value of the wrong type, or a NULL in a field
declared non-nullable.

Fixing the widening rule first is what let the rest collapse rather than
being patched case by case:

- Arithmetic, CASE, COALESCE, GREATEST/LEAST and multi-row VALUES now widen
  across all operands instead of adopting the leftmost. (#121)
- Numeric literals are no longer uniformly int64; a decimal or exponent is
  typed as one. (#122)
- A function of a nullable argument is nullable -- to_date, age, power,
  sqrt, position, LIKE, IN, BETWEEN and comparisons all propagate it now
  rather than hardcoding NOT NULL. (#120, #119)
- EXTRACT, ROUND and generate_series return types come from the dialect
  rather than being assumed PostgreSQL's. (#123)
- The NOT NULL guard matched on an unqualified name, so `WHERE x IS NOT
  NULL` narrowed some other table's `x` in a join. (#124)
- Simple-CASE placeholders are no longer registered as bool parameters.
  (#125)
- array_agg of a paren row-constructor no longer reports the first field's
  type and drops the rest. (#117)

The fixtures are the point as much as the code: #163 recorded that these
shipped because nothing exercised them. 25 new fixtures cover each case,
and two existing ones are corrected.

Closes #117, #119, #120, #121, #122, #123, #124, #125
Refs #163

+++ crates/scythe-core/src/analyzer/mod.rs
@@ -1353,7 +1353,9 @@ WITH t(a, b) AS (SELECT 1, 2) SELECT a, b FROM t;",
 
         let names: Vec<&str> = result.columns.iter().map(|c| c.name.as_str()).collect();
         assert_eq!(names, ["a", "b"]);
-        assert!(result.columns.iter().all(|c| c.neutral_type == "int64"));
+        // `1` and `2` are small integer literals, typed `int32` to match
+        // PostgreSQL (see #122).
+        assert!(result.columns.iter().all(|c| c.neutral_type == "int32"));
     }
 
     /// `SELECT *` over a CTE with an explicit column alias list expands to the
@@ -1414,6 +1416,9 @@ WITH RECURSIVE t(n) AS (SELECT 1 UNION ALL SELECT n + 1 FROM t WHERE n < 10) SEL
 
         let names: Vec<&str> = result.columns.iter().map(|c| c.name.as_str()).collect();
         assert_eq!(names, ["n"]);
-        assert_eq!(result.columns[0].neutral_type, "int64");
+        // `1` and `n + 1` are both `int32` (small integer literal, and the
+        // anchor-typed `n` widened against another `int32` literal); see
+        // #122.
+        assert_eq!(result.columns[0].neutral_type, "int32");
     }
 }
